# Tweak ContinuableScopeManager for better inlining