@richardstartin 2020-07-13
@randomanderson these changes aim to undo a 10% regression since 0.57.0 which became apparent in the reliability environment as soon as the change increasing the size of `ContinuableScopeManager.activate` was merged. It's difficult to provoke in a microbenchmark, and yes, it's very brittle. I'm not sure this adds much complexity really.